Refactored UI code.
