+++ flexmeasures/ui/__init__.py
@@ -44,9 +44,9 @@ flexmeasures_ui = Blueprint(
 def register_at(app: Flask):
     """This can be used to register this blueprint together with other ui-related things"""
 
-    from flexmeasures.ui.crud.assets import AssetCrudUI
-    from flexmeasures.ui.crud.users import UserCrudUI
-    from flexmeasures.ui.crud.accounts import AccountCrudUI
+    from flexmeasures.ui.views.assets import AssetCrudUI
+    from flexmeasures.ui.views.users.users import UserCrudUI
+    from flexmeasures.ui.views.accounts import AccountCrudUI
     from flexmeasures.ui.views.sensors import SensorUI
     from flexmeasures.ui.utils.color_defaults import get_color_settings
