--- flexmeasures/ui/crud/__init__.py
@@ -1,3 +1,0 @@
-"""
-Backoffice UI for CRUD functionality
-"""
