+++ flexmeasures/ui/templates/assets/asset_context.html
