+++ flexmeasures/ui/templates/assets/asset_graph.html
