+++ flexmeasures/ui/templates/assets/asset_new.html
