+++ flexmeasures/ui/templates/assets/asset_properties.html
