+++ flexmeasures/ui/templates/assets/assets.html
