+++ flexmeasures/ui/templates/dashboard.html
