+++ flexmeasures/ui/templates/sensors/sensor_new.html
