+++ flexmeasures/ui/templates/sensors/sensors.html
