+++ flexmeasures/ui/templates/sensors/status.html
