+++ flexmeasures/ui/views/accounts.py
@@ -9,7 +9,7 @@ from flask_security.core import current_user
 
 from flexmeasures.auth.policy import user_has_admin_access, check_access
 
-from flexmeasures.ui.crud.api_wrapper import InternalApi
+from flexmeasures.ui.views.api_wrapper import InternalApi
 from flexmeasures.ui.utils.view_utils import render_flexmeasures_template
 from flexmeasures.data.models.audit_log import AuditLog
 from flexmeasures.data.models.user import Account
@@ -40,7 +40,7 @@ class AccountCrudUI(FlaskView):
         """/accounts"""
 
         return render_flexmeasures_template(
-            "crud/accounts.html",
+            "accounts/accounts.html",
         )
 
     @login_required
@@ -69,7 +69,7 @@ class AccountCrudUI(FlaskView):
             user_can_update_account = False
 
         return render_flexmeasures_template(
-            "crud/account.html",
+            "accounts/account.html",
             account=account,
             accounts=accounts,
             include_inactive=include_inactive,
@@ -87,7 +87,7 @@ class AccountCrudUI(FlaskView):
         audit_logs_response = audit_log_response.json()
 
         return render_flexmeasures_template(
-            "crud/account_audit_log.html",
+            "accounts/account_audit_log.html",
             audit_logs=audit_logs_response,
             account=account,
         )
