+++ flexmeasures/ui/views/assets/__init__.py
@@ -1,0 +1,11 @@
+from flexmeasures.ui.views.assets.forms import AssetForm, NewAssetForm  # noqa: F401
+from flexmeasures.ui.views.assets.utils import (  # noqa: F401
+    get_allowed_price_sensor_data,
+    get_allowed_inflexible_sensor_data,
+    process_internal_api_response,
+    user_can_create_assets,
+    user_can_delete,
+    user_can_update,
+    get_assets_by_account,
+)
+from flexmeasures.ui.views.assets.views import AssetCrudUI  # noqa: F401
