+++ flexmeasures/ui/views/assets/forms.py
